manim-render-service/requirements.txt
manim
supabase
python-dotenv
psycopg2-binary
